@@ -14,42 +14,42 @@
     },
     "devDependencies": {
         "@benrbray/remark-cite": "2.0.1-alpha.4",
-        "@cartamd/plugin-anchor": "^2.0.3",
-        "@graphql-codegen/cli": "^5.0.2",
-        "@graphql-codegen/typed-document-node": "^5.0.9",
-        "@graphql-codegen/typescript": "^4.0.9",
-        "@graphql-codegen/typescript-operations": "^4.2.3",
+        "@cartamd/plugin-anchor": "^2.2.0",
+        "@graphql-codegen/cli": "^5.0.5",
+        "@graphql-codegen/typed-document-node": "^5.1.1",
+        "@graphql-codegen/typescript": "^4.1.6",
+        "@graphql-codegen/typescript-operations": "^4.6.0",
         "@graphql-typed-document-node/core": "^3.2.0",
-        "@microflash/rehype-figure": "^2.1.0",
-        "@sveltejs/adapter-auto": "^3.2.4",
-        "@sveltejs/kit": "^2.5.22",
-        "@sveltejs/vite-plugin-svelte": "^3.1.1",
-        "@types/eslint": "^8.56.11",
-        "@types/node": "^22.5.1",
-        "eslint": "^9.9.0",
+        "@microflash/rehype-figure": "^2.1.4",
+        "@sveltejs/adapter-auto": "^3.3.1",
+        "@sveltejs/kit": "^2.20.7",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0-next-6",
+        "@types/eslint": "^8.56.12",
+        "@types/node": "^22.14.1",
+        "eslint": "^9.24.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.43.0",
-        "globals": "^15.9.0",
+        "eslint-plugin-svelte": "^2.46.1",
+        "globals": "^15.15.0",
         "mdast-util-infobox": "^1.0.5",
-        "prettier": "^3.3.3",
-        "prettier-plugin-svelte": "^3.2.6",
+        "prettier": "^3.5.3",
+        "prettier-plugin-svelte": "^3.3.3",
         "remark-breaks": "^4.0.0",
         "remark-infobox": "^1.0.3",
-        "remark-rehype": "^11.1.0",
+        "remark-rehype": "^11.1.2",
         "remark-toc": "^9.0.0",
         "remark-wiki-link": "^2.0.1",
-        "svelte": "5.0.0-next.224",
-        "svelte-check": "^3.8.5",
-        "tslib": "^2.6.3",
-        "typescript": "^5.5.4",
-        "typescript-eslint": "^8.1.0",
-        "vite": "^5.4.1"
+        "svelte": "5.27.0",
+        "svelte-check": "^3.8.6",
+        "tslib": "^2.8.1",
+        "typescript": "^5.8.3",
+        "typescript-eslint": "^8.30.1",
+        "vite": "^5.4.18"
     },
     "type": "module",
     "dependencies": {
-        "@urql/svelte": "^4.2.1",
+        "@urql/svelte": "^4.2.3",
         "carta-md": "^4.9.0",
-        "graphql": "^16.9.0",
+        "graphql": "^16.10.0",
         "isomorphic-dompurify": "^2.23.0"
     }
 }
