@@ -1,7 +1,9 @@
 import type { LayoutServerLoad } from "./$types";
 
 export const load: LayoutServerLoad = async () => {
+   let googleAnalytics = import.meta.env.VITE_GOOGLE_ANALYTICS_ID
+   
     return {
-        google_analytics: import.meta.env.VITE_GOOGLE_ANALYTICS_ID
+        googleAnalytics
     }
 }
