@@ -1,15 +1,10 @@
-import { client } from "$lib/graphql/client";
-import { TAG } from "$lib/graphql/operations/query";
+import getTagQuery from "$lib/graphql/operations/tags/getTag/getTagQuery";
 import type { PageServerLoad } from "./$types";
 
 export const load: PageServerLoad = async ({params}) => {
-    const tag = await client.query(TAG, {
-        input: {
-            name: params.tag
-        }
-    }).toPromise()
+    const tag = await getTagQuery(params.tag);
 
     return {
-        tag: tag.data.tag.data
+        tag
     }
 }
